Switch to String View Array
Instead of the character array, a standard array of string views are used. This has both the benefit of being much more inline with C++ conventions and the size of the strings are easy to extract to use for generating custom buffer sizes for the pmr strings. It also does away with the reinterpret cast used before.