Using CLDR For Locale Data
Given that currently custom locales will need to be generated in a header format, an up to date complete source of locale data was needed. CLDR seems to fit the criteria while also being in an easy to parse XML format. This should provide a reliable source for generation and the data can be read be easily accessed to be integrated into the project files.