Unique Array Names Added
Started by adding .pot extraction function which takes the comment and msgid from the pot file and stores the map name if one is given in the comment. If a map name is not given, the key string is hashed and the absolute value is taken to achieve a realistically unique name.