Straightforward Way to Access XML Data
After testing multiple ways to access the CLDR XML data in a consistent, locale agnostic way, a series of increasingly specific keys passed in the order they appear in the XML seems best. Further work needs to be done to handle multiple attributes, but this should be a good starting place.